+++ common/src/main/java/com/selfdot/cobblemonmegas/common/CobblemonMegas.java
@@ -12,10 +12,16 @@ import com.selfdot.cobblemonmegas.common.command.CommandTree;
 import com.selfdot.cobblemonmegas.common.item.MegaStoneHeldItemManager;
 import com.selfdot.cobblemonmegas.common.util.DisableableMod;
 import com.selfdot.cobblemonmegas.common.util.MegaUtils;
+import dev.architectury.event.EventResult;
 import dev.architectury.event.events.common.CommandRegistrationEvent;
 import dev.architectury.event.events.common.LifecycleEvent;
+import dev.architectury.event.events.common.PlayerEvent;
 import kotlin.Unit;
+import net.minecraft.entity.ItemEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.network.ServerPlayerEntity;
 
 import java.util.HashSet;
 import java.util.List;
@@ -30,10 +36,15 @@ public class CobblemonMegas extends DisableableMod {
         return INSTANCE;
     }
 
+    private MinecraftServer server;
     private final Set<UUID> TO_MEGA_EVOLVE_THIS_TURN = new HashSet<>();
     private final Set<UUID> HAS_MEGA_EVOLVED_THIS_BATTLE = new HashSet<>();
 
     private Config config;
+
+    public MinecraftServer getServer() {
+        return server;
+    }
 
     public Set<UUID> getToMegaEvolveThisTurn() {
         return TO_MEGA_EVOLVE_THIS_TURN;
@@ -53,6 +64,8 @@ public class CobblemonMegas extends DisableableMod {
         CommandRegistrationEvent.EVENT.register(CommandTree::register);
         LifecycleEvent.SERVER_STARTING.register(this::onServerStarting);
         CobblemonEvents.BATTLE_STARTED_PRE.subscribe(Priority.NORMAL, this::onBattleStartedPre);
+        PlayerEvent.PICKUP_ITEM_POST.register(this::onItemPickup);
+        PlayerEvent.PLAYER_JOIN.register(this::onPlayerJoin);
     }
 
     private static void registerAspectProvider(String aspect) {
@@ -60,6 +73,7 @@ public class CobblemonMegas extends DisableableMod {
     }
 
     private void onServerStarting(MinecraftServer server) {
+        this.server = server;
         registerAspectProvider(DataKeys.MEGA);
         registerAspectProvider(DataKeys.MEGA_X);
         registerAspectProvider(DataKeys.MEGA_Y);
@@ -79,6 +93,14 @@ public class CobblemonMegas extends DisableableMod {
     private Unit onBattleStartedPre(BattleStartedPreEvent event) {
         MegaUtils.deMegaEvolveAll(event.getBattle());
         return Unit.INSTANCE;
+    }
+
+    private void onItemPickup(PlayerEntity player, ItemEntity itemEntity, ItemStack itemStack) {
+        MegaUtils.updateKeyStoneGlow(itemStack, player);
+    }
+
+    private void onPlayerJoin(ServerPlayerEntity player) {
+        MegaUtils.updateKeyStoneGlow(player);
     }
 
 }