+++ common/src/main/java/com/selfdot/cobblemonmegas/common/DataKeys.java
@@ -7,6 +7,7 @@ public class DataKeys {
     public static final String MEGA_X = "mega-x";
     public static final String MEGA_Y = "mega-y";
     public static final String NBT_KEY_MEGA_STONE = "megaStone";
+    public static final String NBT_KEY_KEY_STONE = "keyStone";
     public static final String MEGA_STONE_WHITELIST = "megaStoneWhitelist";
     public static final String MEGA_RAYQUAZA_ALLOWED = "megaRayquazaAllowed";