+++ common/src/main/java/com/selfdot/cobblemonmegas/common/command/MegaEvolveInBattleCommand.java
@@ -22,31 +22,7 @@ public class MegaEvolveInBattleCommand implements Command<ServerCommandSource> {
     public int run(CommandContext<ServerCommandSource> context) {
         ServerPlayerEntity player = context.getSource().getPlayer();
         if (player == null) return 0;
-
-        PokemonBattle battle = BattleRegistry.INSTANCE.getBattleByParticipatingPlayer(player);
-        if (battle == null) {
-            context.getSource().sendError(Text.literal("This can only be used in battle."));
-            return -1;
-        }
-
-        BattleActor playerBattleActor = battle.getActor(player);
-        if (playerBattleActor == null) return 0;
-        List<ActiveBattlePokemon> activeBattlePokemon = playerBattleActor.getActivePokemon();
-        if (activeBattlePokemon.size() != 1) return 0;
-        BattlePokemon battlePokemon = activeBattlePokemon.get(0).getBattlePokemon();
-        if (battlePokemon == null) return 0;
-        Pokemon pokemon = battlePokemon.getEffectedPokemon();
-
-        String reasonCannotMegaEvolve = MegaUtils.reasonCannotMegaEvolve(player, pokemon);
-        if (reasonCannotMegaEvolve != null) {
-            context.getSource().sendError(Text.literal(reasonCannotMegaEvolve));
-            return -1;
-        }
-
-        CobblemonMegas.getInstance().getToMegaEvolveThisTurn().add(playerBattleActor.getUuid());
-        context.getSource().sendMessage(Text.literal(
-            pokemon.getDisplayName().getString() + " will mega evolve this turn if a move is used."
-        ));
+        if (!MegaUtils.attemptMegaEvolveInBattle(player, true)) return -1;
         return SINGLE_SUCCESS;
     }