+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+- Added key stone item which can be used to mega evolve in battle.
+- Added /givekeystone \<player\> (selfdot.megas.givekeystone)
+
 # 1.0.2
 
 - Added config option "megaRayquazaAllowed" to control whether mega rayquaza can be used.